+++ CHANGELOG.md
@@ -21,7 +21,7 @@ For details about compatibility between different releases, see the **Commitment
 
 ### Security
 
-## [3.35.2] - unreleased
+## [3.35.2] - 2026-01-30
 
 ### Changed